requirements.txt
aiohttp==3.9.5
aiokafka==0.10.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
av==12.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.4.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
environs==11.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-offline==1.7.1
fastapi_cors==0.0.6
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
opencv-python-headless==4.10.0.82
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.2
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
